+++ ecs/bitmask_test.go
@@ -41,3 +41,27 @@ func TestBitMask(t *testing.T) {
 	assert.True(t, mask.ContainsAny(other1))
 	assert.False(t, mask.ContainsAny(other2))
 }
+
+func TestBitMask128(t *testing.T) {
+	for i := 0; i < MaskTotalBits; i++ {
+		mask := NewBitMask(ID(i))
+		assert.Equal(t, 1, mask.TotalBitsSet())
+		assert.True(t, mask.Get(ID(i)))
+	}
+	mask := BitMask{}
+	assert.Equal(t, 0, mask.TotalBitsSet())
+
+	for i := 0; i < MaskTotalBits; i++ {
+		mask.Set(ID(i), true)
+		assert.Equal(t, i+1, mask.TotalBitsSet())
+		assert.True(t, mask.Get(ID(i)))
+	}
+
+	mask = NewBitMask(ID(1), ID(2), ID(13), ID(27), ID(63), ID(64), ID(65))
+
+	assert.True(t, mask.Contains(NewBitMask(ID(1), ID(2), ID(63), ID(64))))
+	assert.False(t, mask.Contains(NewBitMask(ID(1), ID(2), ID(63), ID(90))))
+
+	assert.True(t, mask.ContainsAny(NewBitMask(ID(6), ID(65), ID(111))))
+	assert.False(t, mask.ContainsAny(NewBitMask(ID(6), ID(66), ID(90))))
+}